Move `ExpressionSerializer` back to `:opensearch`.
